Fix int64 literal truncation, finally step starvation; cover debug hook
- Integer literals no longer truncate to int32: the parser keeps values
  past int32 as int64, so `return 9223372036854775807` and `-5000000000`
  evaluate correctly (previously both silently wrapped to int32).
- Decimal literals parse via stoull like the base-prefixed forms, so
  -9223372036854775808 (INT64_MIN) is now writable; values in [2^63, 2^64)
  wrap into two's complement, consistent with 0x/0b/0o literals.
- Unary minus negates in the unsigned domain — no signed-overflow UB on
  INT64_MIN — and preserves int64 width for 64-bit operands. Same wrap
  applied to enum member negation.
- `finally` blocks get a bounded 128-step reserve beyond an exhausted step
  budget so cleanup completes on fatal step-limit errors; a finally that
  outspends the reserve still dies (guardrail intact).
- Tests: re-entrant executeProcedure exercising the ScriptException
  "Uncaught exception" path, DebugContext hook coverage (procedure, file,
  line, callStack, visible variables), literal overflow wrap.

+++ src/Lexer.cpp
@@ -407,7 +407,10 @@ Token Lexer::number() {
     if (isFloat) {
       token.doubleValue = std::stod(numStr);
     } else {
-      token.intValue = std::stoll(numStr);
+      // Parse as unsigned so values up to 2^64-1 lex like the base-prefixed
+      // forms: they wrap into two's-complement int64, which also makes
+      // -9223372036854775808 (INT64_MIN) writable as a literal.
+      token.intValue = static_cast<int64_t>(std::stoull(numStr));
     }
   } catch (const std::exception &) {
     token.type = TokenType::UNKNOWN;
